#12 JankenTextのModel化
