+++ CheatingJanken/CheatingJanken/ViewModel/JankenTextViewModel.swift
@@ -7,19 +7,12 @@
 
 import Foundation
 
-struct JankenTextViewModel {
-    func jankenText(jankenCount: Int) -> String {
-        var jankenText: String = ""
-        // カウント毎にテキストを変更する
-        switch jankenCount {
-        case 0, 1, 2: jankenText = "Ready???"
-        case 3: jankenText = "最初は、、"
-        case 4: jankenText = "ぐー！"
-        case 5: jankenText = "じゃんけん"
-        case 6, 7: jankenText = "ぽん！！！"
-        default: break
-        }
-
-        return jankenText
+class JankenTextViewModel: ObservableObject {
+    // JankenTextModelのインスタンス生成
+    let jankenTextModel = JankenTextModel()
+    // jankenTextをPublish
+    @Published var jankenText = ""
+    func makeJankenText(jankenCount: Int) {
+        jankenText = jankenTextModel.jankenText(jankenCount: jankenCount)
     }
 }
